+++ src/test/java/hu/martin/chatter/adapter/in/web/conversation/StubConversationServiceForSendingMessage.java
@@ -1,13 +1,11 @@
 package hu.martin.chatter.adapter.in.web.conversation;
 
 import hu.martin.chatter.application.ConversationService;
-import hu.martin.chatter.domain.Conversation;
-import hu.martin.chatter.domain.ConversationId;
-import hu.martin.chatter.domain.Message;
-import hu.martin.chatter.domain.MessageId;
-import hu.martin.chatter.domain.ParticipantId;
+import hu.martin.chatter.domain.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+
+import java.math.BigInteger;
 
 public class StubConversationServiceForSendingMessage implements ConversationService {
 
@@ -42,7 +40,7 @@ public class StubConversationServiceForSendingMessage implements ConversationSer
 
   @Override
   public Mono<Message> receiveMessage(Message message) {
-    message.setId(MessageId.of(1L));
+    message.setId(MessageId.of(BigInteger.valueOf(1L)));
     return Mono.just(message);
   }
