+++ src/test/java/hu/martin/chatter/domain/ConversationFactory.java
@@ -1,5 +1,6 @@
 package hu.martin.chatter.domain;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 
 public class ConversationFactory {
@@ -12,7 +13,7 @@ public class ConversationFactory {
 
   public static Conversation withDefaults() {
     Conversation conversation = new Conversation();
-    conversation.setId(ConversationId.of(1L));
+    conversation.setId(ConversationId.of(BigInteger.valueOf(1L)));
     return conversation;
   }
 }
