+++ src/test/java/hu/martin/chatter/domain/MessageFactory.java
@@ -1,11 +1,12 @@
 package hu.martin.chatter.domain;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public class MessageFactory {
 
   public static Message withDefaults() {
-    ParticipantId participantId = ParticipantId.of(1L);
+    ParticipantId participantId = ParticipantId.of(BigInteger.valueOf(1L));
     return defaultWithSender(participantId);
   }
 
@@ -19,15 +20,15 @@ public class MessageFactory {
     MessageContent messageContent = MessageContent.of("");
     CreatedDateTime createdDateTime = CreatedDateTime.of(LocalDateTime.now());
     Message message = new Message(participantId, messageContent, createdDateTime);
-    message.setId(MessageId.of(1L));
+    message.setId(MessageId.of(BigInteger.valueOf(1L)));
     return message;
   }
 
   public static Message defaultWIthCreatedDateTimeOf(CreatedDateTime createdDateTime) {
-    ParticipantId participantId = ParticipantId.of(1L);
+    ParticipantId participantId = ParticipantId.of(BigInteger.valueOf(1L));
     MessageContent messageContent = MessageContent.of("");
     Message message = new Message(participantId, messageContent, createdDateTime);
-    message.setId(MessageId.of(1L));
+    message.setId(MessageId.of(BigInteger.valueOf(1L)));
     return message;
   }
